bug with event listener still there
